+++ src/alembic/env.py
@@ -1,12 +1,11 @@
-from app.database import Base
-from app import models
 from logging.config import fileConfig
 
-from sqlalchemy import engine_from_config, create_engine
+from app.database import Base
+from app import models
+
+from sqlalchemy import engine_from_config
 from sqlalchemy import pool
-
 from alembic import context
-
 from app.settings import cfg
 
 # this is the Alembic Config object, which provides
@@ -19,6 +18,8 @@ fileConfig(config.config_file_name)
 
 # add your model's MetaData object here
 # for 'autogenerate' support
+# from myapp import mymodel
+# target_metadata = mymodel.Base.metadata
 target_metadata = Base.metadata
 
 # other values from the config, defined by the needs of env.py,
@@ -39,7 +40,7 @@ def run_migrations_offline():
     script output.
 
     """
-    url = get_db_url()
+    url = cfg.database_url
     context.configure(
         url=url,
         target_metadata=target_metadata,
@@ -58,13 +59,13 @@ def run_migrations_online():
     and associate a connection with the context.
 
     """
+
     from app.database import engine
+
     connectable = engine
 
     with connectable.connect() as connection:
-        context.configure(
-            connection=connection, target_metadata=target_metadata
-        )
+        context.configure(connection=connection, target_metadata=target_metadata)
 
         with context.begin_transaction():
             context.run_migrations()
