+++ src/alembic/versions/4ee595c9eacd_.py
@@ -1,0 +1,24 @@
+"""empty message
+
+Revision ID: 4ee595c9eacd
+Revises: 
+Create Date: 2021-08-29 14:31:45.776135
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '4ee595c9eacd'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    pass
+
+
+def downgrade():
+    pass
