--- src/alembic/versions/dd77d862eee2_initial_tables.py
@@ -1,53 +1,0 @@
-"""Initial Tables
-
-Revision ID: dd77d862eee2
-Revises: 
-Create Date: 2021-07-30 01:29:54.955518
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'dd77d862eee2'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('user',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=64), nullable=False),
-    sa.Column('password_hash', sa.String(length=128), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('is_superuser', sa.Boolean(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_index(op.f('ix_user_is_superuser'), 'user', ['is_superuser'], unique=False)
-    op.create_index(op.f('ix_user_name'), 'user', ['name'], unique=True)
-    op.create_index(op.f('ix_user_password_hash'), 'user', ['password_hash'], unique=False)
-    op.create_table('todo',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('title', sa.String(length=128), nullable=False),
-    sa.Column('notes', sa.String(length=256), nullable=False),
-    sa.Column('completed', sa.Boolean(), nullable=False),
-    sa.Column('owner_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['owner_id'], ['user.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('todo')
-    op.drop_index(op.f('ix_user_password_hash'), table_name='user')
-    op.drop_index(op.f('ix_user_name'), table_name='user')
-    op.drop_index(op.f('ix_user_is_superuser'), table_name='user')
-    op.drop_table('user')
-    # ### end Alembic commands ###
