--- src/app/tests/unit/test_auth.py
@@ -1,12 +1,0 @@
-import unittest
-
-
-def test_verify_pw():
-    pass
-
-
-def test_get_pw_hash():
-
-
-if __name__ == '__main__':
-    unittest.main()
